API testing, changed RefType in `RecursiveObj<RefType>` to AT
Since each `Enum<RefType>` is of a specific reference type, it doesn't
need to be able to implement both `RecursiveObj<A>` and
`RecursiveObj<B>` at the same time.  The single implementation with an
associated type allows the same restrictions on generics, without
requiring the user to specify the generic trait being fulfilled.