API testing, added lifetime parameter to recursive family/type/obj
In order to have the VisitingRef hold a `&'a T` reference to
non-recursive objects instead of a copied `T`, need to have somewhere
that the `'a` lifetime will be propagated through.  Unfortunately,
couldn't find a way to avoid this, and it makes some of the
declarations (e.g. `impl TypedTree`) be weirdly verbose.