Pulled <'a> into type of RecursiveFamily, TryAsRef trait
I think this starts getting to a cleaner API.  The `Container:
TryAsRef<Node>` trait would only be needed to visit a graph, and the
`Container: From<Node>` trait would only be needed to build a graph.

Alternatively, I might just be going around in circles, and I still
don't have a way to avoid the repetition of `'a` in `Node<'a, 'a,
Storage>`.