Finally removed unnecessary lifetime parameter
Previously, I hadn't been able to keep the view-ing references
contained in the `Visiting::Ref` associated type, because there needed
to be a way to indicate that `&'view T` was a valid type.  I had
previously tried making `trait RecursiveRefType` take a lifetime
parameter, but that still required that view-ing lifetime to be known
at the use site.

The trick was to implement `RecursiveRefType<'ext>` for all
`Visiting<'view>` where `'ext: 'view`.  That is, implementing the
trait for all lifetimes larger than the view.  Then, the trait can
require that any argument to `RecursiveRefType::Ref` and
`RecursiveRefType::Value` outlive `'ext`, and is therefore safe to
make a reference of lifetime `'view`.

Before pulling this back into the auto-generating macros, I'll want to
see if this improves the container situation.  I initially moved to
having a single defined container for each family because it was
tedious to specify in all cases, including additional lifetime
bounds.  However, with the view-ing lifetime no longer exposed, I may
be able to change `Visiting` to keep track of the container as well,
without requiring the user to specify it everywhere.