Updated PartialEq to remove wildcard match
When matching on `(self, other)`, there's a wildcard `_ => false` to
catch mismatched types.  This wildcard also makes the match
exhaustive, even if a new node type has been implemented.  Rewriting
to use `match self`, with `matches!(other, ...)` in the value, so that
the addition of new operations causes a compilation error, and
requiring these implementations to be updated.