Added *::container::* trait alias (e.g. expr::container::IntExpr)
Since an `Arena` is generic over the container type that it uses, it
can be useful to specify "An Arena whose container is sufficient to
hold type T or any of its referents".  The auto-generated
`mod_name::container::enum_name` traits imply and are implemented for
the `TryAsRef` and `From` implementations that are required for the
`enum_name`.